+++ crates/nix_rs/src/health/check/caches.rs
@@ -2,7 +2,7 @@ use leptos::*;
 use serde::{Deserialize, Serialize};
 use url::Url;
 
-use crate::nix::{
+use crate::{
     config::ConfigVal,
     health::{
         report::{Report, WithDetails},
@@ -11,7 +11,7 @@ use crate::nix::{
     info,
 };
 
-/// Check that [crate::nix::config::NixConfig::substituters] is set to a good value.
+/// Check that [crate::config::NixConfig::substituters] is set to a good value.
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Caches(ConfigVal<Vec<Url>>);
