+++ crates/nix_rs/src/health/report.rs
