+++ crates/nix_rs/src/health/traits.rs
