+++ crates/nix_rs/src/info.rs
