+++ crates/nix_rs/src/lib.rs
@@ -1,3 +1,9 @@
 #![feature(associated_type_defaults)]
-mod command;
-pub mod nix;
+//! Rust module to interact with Nix
+pub mod command;
+pub mod config;
+pub mod flake;
+pub mod health;
+pub mod info;
+pub mod refs;
+pub mod version;
