--- crates/nix_rs/src/nix/mod.rs
@@ -1,8 +1,0 @@
-//! Rust module to interact with Nix
-pub mod command;
-pub mod config;
-pub mod flake;
-pub mod health;
-pub mod info;
-pub mod refs;
-pub mod version;
